did helpers: Move url_to_did_web from tests
Helpful for OIDC audience
